final project

+++ parallelHillClimber.py
@@ -2,19 +2,27 @@ import os
 from solution import SOLUTION
 import constants as c
 import copy
+import pickle
 
 class PARALLEL_HILLCLIMBER:
-    def __init__(self):
+    def __init__(self, pickle_file):
 
         os.system("rm brain*.nndf")
         os.system("rm fitness*.txt")
         os.system("rm body*.urdf")
+        self.pickle_file = pickle_file
+        self.seed = c.seed
         # os.system("rm allBestFitness0.txt")
+
+        self.listBestFitnessSoFar = []
+        self.listBestFitnessSoFarSolutions = []
+        self.bestfitnessSoFar = float('inf')
+        self.currentGeneration = 0
 
         self.parents = {}
         self.nextAvailableID = 0
         self.children = {}
-
+        
         for i in range(c.populationSize):
             self.parents[i] = SOLUTION(self.nextAvailableID)
             self.nextAvailableID += 1
@@ -22,21 +30,20 @@ class PARALLEL_HILLCLIMBER:
 
     def Evolve(self):
         self.Evaluate(self.parents)
-        for currentGeneration in range(c.numberOfGenerations):
+        while self.currentGeneration < c.numberOfGenerations:
             self.Evolve_For_One_Generation()
+            self.currentGeneration += 1
+            
 
     def Evolve_For_One_Generation(self):
         self.Spawn()
         self.Mutate()
         self.Evaluate(self.children)
-        # self.child.Evaluate('DIRECT')
         self.Print()
         self.Select()
 
     def Spawn(self):
         self.children = {}
-
-        # self.nextAvailableID += 1
         for parent in self.parents:
             self.children[parent] = copy.deepcopy(self.parents[parent])
             self.children[parent].Set_ID(self.nextAvailableID)
@@ -52,21 +59,42 @@ class PARALLEL_HILLCLIMBER:
 
     def Select(self):
         for parent in self.parents:
-            # print("parent fitness: " +  str(self.parents[parent].fitness) + ", children fitness: " + str(self.children[parent].fitness))
+            # work on saving robots of each best generation 
+
+            ### if child is better than parent. 
             if (self.parents[parent].fitness > self.children[parent].fitness):
-                os.system("rm brain" + str(self.parents[parent].myID) + ".nndf")
-                bodyFileName = "body" + str(self.parents[parent].myID) + ".urdf"
-                os.system("rm " + bodyFileName)
+                ## If parent used to be a best in generation, keep. Else, delete.
+                if (self.parents[parent].myID not in self.listBestFitnessSoFar):
+                    bodyFileName = "body" + str(self.parents[parent].myID) + ".urdf"
+                    os.system("rm brain" + str(self.parents[parent].myID) + ".nndf")
+                    os.system("rm " + bodyFileName)
                 self.parents[parent] = self.children[parent]
-            else:
+            ## if child is worse, don't keep.
+            else: 
                 os.system("rm brain" + str(self.children[parent].myID) + ".nndf")
                 bodyFileName = "body" + str(self.children[parent].myID) + ".urdf"
+                # if (self.children[parent].myID not in self.listBestFitnessSoFar):
                 os.system("rm " + bodyFileName)
         
+        # find best solution in generation and record fitness
         bestfit = float('inf')
+        bestSolution = 0
         for parent in self.parents:
             if self.parents[parent].fitness < bestfit:
                 bestfit = self.parents[parent].fitness
+                bestSolution = self.parents[parent]
+        
+        # if best solution in generation is better than best solution so far, record this down. (pickled, too)
+        if (bestfit < self.bestfitnessSoFar):
+            new_pickle_file = "saved_seed" + str(c.seed) + "_gen" + str(self.currentGeneration) + ".pickle"
+            with open(new_pickle_file, 'wb') as f:
+                pickle.dump(self, f)
+            self.bestfitnessSoFar = bestfit
+            self.listBestFitnessSoFar.append(bestSolution.myID)
+            f = open("increasesInFitness" + str(c.seed) + ".txt", "a")
+            f.write(str(self.bestfitnessSoFar) + " " + str(bestSolution.myID) + '\n')
+            f.close()
+
         f = open("allBestFitness" + str(c.seed) + ".txt", "a")
         f.write(str(bestfit) + '\n')
         f.close()
@@ -84,7 +112,6 @@ class PARALLEL_HILLCLIMBER:
         for parent in self.parents:
             if self.parents[parent].fitness < bestFitness:
                 bestFitness = self.parents[parent].fitness
-                print("FITNESS WTF: " + str(bestFitness))
                 bestParent = self.parents[parent]
         print("Best fitness-  "+ str(bestParent.myID) + ": " + str(bestFitness))
         bestParent.Start_Simulation('GUI')
